Ciemss calibrate initial UI refactor
